requirements.txt
numpy = "*"
pandas = "*"
pillow = "*"
matplotlib = "*"
opencv-python = "*"
albumentations = "*"
tqdm = "*"
pyyaml = "*"
scikit-learn = "*"
wandb = "*"
tensorboard = "*"
aiofiles = "*"
torch = "*"
torchvision = "*"
torchaudio = "*"